webhook_v2/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
requests>=2.31.0

# Database
psycopg[binary]>=3.1.0

# Configuration
pydantic-settings>=2.1.0

# AI/Classification
google-generativeai>=0.4.0

# Logging
structlog>=24.1.0
orjson>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Storage
minio>=7.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
